bug fix: bad malloc

+++ src/generators/rng/gaussian.c
@@ -40,7 +40,7 @@ double *gaussian(int nclasses, double skewness)
 	assert(skewness > 0.0);
 	assert(skewness < 1.0);
 
-	histogram = smalloc(nclasses*sizeof(int));
+	histogram = smalloc(nclasses*sizeof(double));
 
 	/* Build histogram. */
 	freq = 1.0;
